backend/requirements.txt
alembic==1.4.3
attrs==20.2.0
base58==2.0.1
bitarray==1.2.2
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cytoolz==0.11.0
eth-abi==2.1.1
eth-account==0.5.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-API==2.0
Flask-Cors==3.0.9
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
hexbytes==0.2.1
idna==2.10
ipfshttpclient==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
PyJWT==1.7.1
lru-dict==1.1.6
Mako==1.1.3
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
protobuf==3.13.0
psycopg2-binary==2.8.6
pycparser==2.20
pycryptodome==3.9.8
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
requests==2.24.0
rlp==2.0.0a1
rusty-rlp==0.1.15
six==1.15.0
SQLAlchemy==1.3.19
SQLAlchemy-Utils==0.36.8
toolz==0.11.1
urllib3==1.25.10
varint==1.0.2
web3==5.12.1
websockets==8.1
Werkzeug==1.0.1
